+++ mainwindow.cpp
@@ -1,31 +1,15 @@
 ﻿#include "mainwindow.h"
 #include "ui_mainwindow.h"
-#include <QDesktopWidget>
-#include <QNetworkProxy>
-#include "cookiejar.h"
-#include <QNetworkDiskCache>
-#include <QStandardPaths>
-#include "qnetworkproxyfactoryset.h"
-#include "KQWebPage.h"
+
 #include "kandataconnector.h"
-#include <QWinTaskbarProgress>
-#include <QMessageBox>
-#include "kqnetworkaccessmanager.h"
-#include <QNetworkRequest>
-#include <QPixmap>
+
 //#include <QWebFrame>
 #include <QShortcut>
 #include <QWebEngineSettings>
 
-#include <Audiopolicy.h>
-#include <Mmdeviceapi.h>
-#include <QHttpPart>
-
-#include <QAxWidget>
-#include <QSettings>
-#include <QSysInfo>
-#include "ControlManager.h"
-#include "ExpeditionManager.h"
+#ifdef Q_OS_WIN
+#include <windows.h>
+#endif
 
 #define SAFE_RELEASE(x) if(x) { x->Release(); x = NULL; } 
 
@@ -46,7 +30,7 @@ MainWindow::MainWindow(QWidget *parent) :
 	ui->setupUi(this);
 	ui->retranslateUi(this);
 
-	if (_bUseIE)
+	if (_webWidgetType == WebWidgetType::IE)
 	{
 		setAttribute(Qt::WA_TranslucentBackground, false);
 
@@ -156,377 +140,9 @@ MainWindow::~MainWindow()
 	delete ui;
 }
 
-void MainWindow::postInit(InfoMainWindow *pInfo, TimerMainWindow *pTimer, WeaponMainWindow* pWeapon, ShipMainWindow* pShip)
-{
-	_pInfoWindow = pInfo;
-	_pTimerWindow = pTimer;
-	_pWeaponWindow = pWeapon;
-	_pShipWindow = pShip;
-
-	QObject::connect(this, SIGNAL(sigActivated(QWidget*, bool)), _pInfoWindow, SLOT(slotActivate(QWidget*, bool)));
-	QObject::connect(this, SIGNAL(sigActivated(QWidget*, bool)), _pTimerWindow, SLOT(slotActivate(QWidget*, bool)));
-	QObject::connect(this, SIGNAL(sigActivated(QWidget*, bool)), _pWeaponWindow, SLOT(slotActivate(QWidget*, bool)));
-	QObject::connect(this, SIGNAL(sigActivated(QWidget*, bool)), _pShipWindow, SLOT(slotActivate(QWidget*, bool)));
-
-	QObject::connect(_pInfoWindow, SIGNAL(sigActivated(QWidget*, bool)), this, SLOT(slotActivate(QWidget*, bool)));
-	QObject::connect(_pInfoWindow, SIGNAL(sigActivated(QWidget*, bool)), _pTimerWindow, SLOT(slotActivate(QWidget*, bool)));
-	QObject::connect(_pInfoWindow, SIGNAL(sigActivated(QWidget*, bool)), _pWeaponWindow, SLOT(slotActivate(QWidget*, bool)));
-	QObject::connect(_pInfoWindow, SIGNAL(sigActivated(QWidget*, bool)), _pShipWindow, SLOT(slotActivate(QWidget*, bool)));
-
-	QObject::connect(_pTimerWindow, SIGNAL(sigActivated(QWidget*, bool)), _pInfoWindow, SLOT(slotActivate(QWidget*, bool)));
-	QObject::connect(_pTimerWindow, SIGNAL(sigActivated(QWidget*, bool)), _pWeaponWindow, SLOT(slotActivate(QWidget*, bool)));
-	QObject::connect(_pTimerWindow, SIGNAL(sigActivated(QWidget*, bool)), _pShipWindow, SLOT(slotActivate(QWidget*, bool)));
-	QObject::connect(_pTimerWindow, SIGNAL(sigActivated(QWidget*, bool)), this, SLOT(slotActivate(QWidget*, bool)));
-
-	QObject::connect(_pWeaponWindow, SIGNAL(sigActivated(QWidget*, bool)), _pInfoWindow, SLOT(slotActivate(QWidget*, bool)));
-	QObject::connect(_pWeaponWindow, SIGNAL(sigActivated(QWidget*, bool)), _pTimerWindow, SLOT(slotActivate(QWidget*, bool)));
-	QObject::connect(_pWeaponWindow, SIGNAL(sigActivated(QWidget*, bool)), this, SLOT(slotActivate(QWidget*, bool)));
-	QObject::connect(_pWeaponWindow, SIGNAL(sigActivated(QWidget*, bool)), _pShipWindow, SLOT(slotActivate(QWidget*, bool)));
-
-	QObject::connect(_pShipWindow, SIGNAL(sigActivated(QWidget*, bool)), _pTimerWindow, SLOT(slotActivate(QWidget*, bool)));
-	QObject::connect(_pShipWindow, SIGNAL(sigActivated(QWidget*, bool)), this, SLOT(slotActivate(QWidget*, bool)));
-	QObject::connect(_pShipWindow, SIGNAL(sigActivated(QWidget*, bool)), _pWeaponWindow, SLOT(slotActivate(QWidget*, bool)));
-	QObject::connect(_pShipWindow, SIGNAL(sigActivated(QWidget*, bool)), _pInfoWindow, SLOT(slotActivate(QWidget*, bool)));
-
-    connect(_pInfoWindow, SIGNAL(sigRestoreMinimizeToggled(bool)), this, SLOT(slotToggleRestoreMinimize(bool)));
-	connect(_pTimerWindow, SIGNAL(sigRestoreMinimizeToggled(bool)), this, SLOT(slotToggleRestoreMinimize(bool)));
-	connect(_pWeaponWindow, SIGNAL(sigRestoreMinimizeToggled(bool)), this, SLOT(slotToggleRestoreMinimize(bool)));
-	connect(_pShipWindow, SIGNAL(sigRestoreMinimizeToggled(bool)), this, SLOT(slotToggleRestoreMinimize(bool)));
-	connect(this, SIGNAL(sigRestoreMinimizeToggled(bool)), this, SLOT(slotSelfToggleRestoreMinimize(bool)));
-
-
-	connect(ui->pbCheckInfo, SIGNAL(toggled(bool)), _pInfoWindow, SLOT(setVisible(bool)));
-	connect(ui->pbCheckTimer, SIGNAL(toggled(bool)), _pTimerWindow, SLOT(setVisible(bool)));
-    connect(ui->pbCheckWeapon, SIGNAL(toggled(bool)), _pWeaponWindow, SLOT(slotSetVisible(bool)));
-	connect(ui->pbCheckShip, SIGNAL(toggled(bool)), _pShipWindow, SLOT(slotSetVisible(bool)));
-
-	connect(ui->pbCheckTransparent, SIGNAL(toggled(bool)), this, SLOT(on_pbCheckTransparent_toggled(bool)));
-
-	AdjustVolume(-1);
-
-	HWND hwnd = (HWND)effectiveWinId();
-	RegisterPowerSettingNotification(hwnd, &GUID_CONSOLE_DISPLAY_STATE, DEVICE_NOTIFY_WINDOW_HANDLE);
-	RegisterPowerSettingNotification(hwnd, &GUID_SYSTEM_AWAYMODE, DEVICE_NOTIFY_WINDOW_HANDLE);
-}
-
-void MainWindow::onSubMainWindowShowHide(bool bShow, MainWindowBase *pWindow)
-{
-	QPushButton * pButton = 0;
-	if (pWindow == _pInfoWindow)
-	{
-		pButton = ui->pbCheckInfo;
-	}
-	else if (pWindow == _pTimerWindow)
-	{
-		pButton = ui->pbCheckTimer;
-	}
-	else if (pWindow == _pWeaponWindow)
-	{
-		pButton = ui->pbCheckWeapon;
-	}
-	else if (pWindow == _pShipWindow)
-	{
-		pButton = ui->pbCheckShip;
-	}
-	if (!pButton)
-	{
-		return;
-	}
-
-	pButton->setChecked(bShow);
-}
-
-
-void MainWindow::changeEvent(QEvent *e)
-{
-	MainWindowBase::changeEvent(e);
-
-	if (e->type() == QEvent::WindowStateChange)
-	{
-		QWindowStateChangeEvent* event = static_cast<QWindowStateChangeEvent*>(e);
-		if(event->oldState() & Qt::WindowMinimized)
-		{
-//            if (ui->pbCheckInfo->isChecked())
-            {
-                _pInfoWindow->setWindowState(Qt::WindowNoState);
-                _pInfoWindow->show();
-                _pInfoWindow->stackUnder(this);
-            }
-
-//            if (ui->pbCheckTimer->isChecked())
-            {
-                _pTimerWindow->setWindowState(Qt::WindowNoState);
-                _pTimerWindow->show();
-                _pTimerWindow->stackUnder(_pInfoWindow);
-            }
-
-//            if (ui->pbCheckWeapon->isChecked())
-//            {
-//                m_pWeaponWindow->setWindowState(Qt::WindowNoState);
-//                m_pWeaponWindow->show();
-//                m_pWeaponWindow->stackUnder(m_pWeaponWindow);
-//            }
-			this->raise();
-			this->activateWindow();
-			this->setSleepMode(false);
-		}
-		else if (isMinimized())
-		{
-			if (!_pInfoWindow->isDockingOnTop())
-			{
-				_pInfoWindow->hide();
-			}
-			if (!_pTimerWindow->isDockingOnTop())
-			{
-				_pTimerWindow->hide();
-			}
-			if (!_pWeaponWindow->isDockingOnTop())
-			{
-				_pWeaponWindow->hide();
-			}
-			if (!_pShipWindow->isDockingOnTop())
-			{
-				_pShipWindow->hide();
-			}
-//			this->setSleepMode(true);
-		}
-	}
-
-}
-
-void MainWindow::closeEvent(QCloseEvent *e)
-{
-	if (!ControlManager::getInstance().isActiveRunning())
-	{
-		QMessageBox * pMessageBox = new QMessageBox(
-			QMessageBox::NoIcon
-			, QString::fromLocal8Bit("")
-			, QString::fromLocal8Bit("本当に終了しますか？")
-			, QMessageBox::Yes | QMessageBox::No
-			, this
-			, Qt::Dialog | Qt::MSWindowsFixedSizeDialogHint | Qt::FramelessWindowHint);
-		pMessageBox->setDefaultButton(QMessageBox::No);
-		pMessageBox->setAttribute(Qt::WA_DeleteOnClose, true);
-		int reply = pMessageBox->exec();
-
-		if (reply == QMessageBox::No)
-		{
-			e->ignore();
-			return;
-		}
-	}
-
-	if (_proxyMode == ProxyMode::Fid && _pFid)
-	{
-		_pFid->Shutdown();
-	}
-	else if (_proxyMode == ProxyMode::Nekoxy && _pNekoxy)
-	{
-		_pNekoxy->Shutdown();
-	}
-	else if (_proxyMode == ProxyMode::Titanium && _pTitanium)
-	{
-		_pTitanium->Shutdown();
-	}
-
-	ControlManager::getInstance().Terminate();
-
-	MainWindowBase::closeEvent(e);
-
-	_pInfoWindow->close();
-	_pTimerWindow->close();
-    _pWeaponWindow->close();
-	_pShipWindow->close();
-}
-
-void MainWindow::moveEvent(QMoveEvent *e)
-{
-	MainWindowBase::moveEvent(e);
-
-	QPoint diffPos = e->pos()-e->oldPos();
-
-	if (_bMoveSubTogether)
-	{
-		if (!_pInfoWindow->isDockingOnTop())
-		{
-			_pInfoWindow->move(_pInfoWindow->pos() + diffPos);
-		}
-		if (!_pTimerWindow->isDockingOnTop())
-		{
-			_pTimerWindow->move(_pTimerWindow->pos() + diffPos);
-		}
-		if (!_pWeaponWindow->isDockingOnTop())
-		{
-			_pWeaponWindow->move(_pWeaponWindow->pos() + diffPos);
-		}
-		if (!_pShipWindow->isDockingOnTop())
-		{
-			_pShipWindow->move(_pShipWindow->pos() + diffPos);
-		}
-	}
-}
-
-void MainWindow::mouseReleaseEvent(QMouseEvent *event)
-{
-	QRect rc = QApplication::desktop()->availableGeometry(QApplication::desktop()->primaryScreen());
-	if (x() < 20 && x() > -20 && x()!=0)
-	{
-		move(0, y());
-	}
-	else if (x()+width() > rc.width()-20 && x()+width() < rc.width()+20 && x()!=rc.width()-width())
-	{
-		move(rc.width()-width(), y());
-	}
-	if (y() < 20 && y()!=0)
-	{
-		move(x(), 0);
-	}
-	MainWindowBase::mouseReleaseEvent(event);
-}
-
-void MainWindow::slotActivate(QWidget *w, bool bActivate)
-{
-	Q_UNUSED(bActivate);
-	this->raise();
-	w->stackUnder(this);
-}
-
-void MainWindow::slotToggleRestoreMinimize(bool bRestore)
-{
-	if (bRestore)
-	{
-		if (isMinimized())
-		{
-			this->setWindowState(Qt::WindowNoState);
-		}
-	}
-	if (!bRestore)
-	{
-		if (!isMinimized())
-		{
-			this->setWindowState(Qt::WindowMinimized);
-		}
-	}
-}
-
-void MainWindow::slotSelfToggleRestoreMinimize(bool bRestore)
-{
-	if (!bRestore)
-	{
-		this->AdjustVolume(0);
-	}
-	else
-	{
-		this->AdjustVolume(-1);
-	}
-}
-
 void MainWindow::slotParse(const QString &PathAndQuery, const QString &requestBody, const QString &responseBody)
 {
 	KanDataConnector::getInstance().Parse(PathAndQuery, requestBody, responseBody);
-}
-
-void MainWindow::on_pbClose_clicked()
-{
-	close();
-}
-
-void MainWindow::on_pbMinimize_clicked()
-{
-	setWindowState(Qt::WindowMinimized);
-}
-
-void MainWindow::on_pbCheckTransparent_toggled(bool checked)
-{
-	static QList<QPair<QString, QWidget*>> s_listpair;
-	if (checked)
-	{
-		if (!_bUseIE)
-		{
-			s_listpair.clear();
-
-			QList<QWidget *> lstWidget = findChildren<QWidget *>();
-			QWidget * w;
-			foreach(w, lstWidget)
-			{
-				QPair<QString, QWidget*> pair;
-				pair.first = w->styleSheet();
-				pair.second = w;
-				s_listpair.append(pair);
-			}
-
-			lstWidget = _pInfoWindow->findChildren<QWidget *>();
-			foreach(w, lstWidget)
-			{
-				QPair<QString, QWidget*> pair;
-				pair.first = w->styleSheet();
-				pair.second = w;
-				s_listpair.append(pair);
-			}
-			lstWidget = _pTimerWindow->findChildren<QWidget *>();
-			foreach(w, lstWidget)
-			{
-				QPair<QString, QWidget*> pair;
-				pair.first = w->styleSheet();
-				pair.second = w;
-				s_listpair.append(pair);
-			}
-			lstWidget = _pWeaponWindow->findChildren<QWidget *>();
-			foreach(w, lstWidget)
-			{
-				QPair<QString, QWidget*> pair;
-				pair.first = w->styleSheet();
-				pair.second = w;
-				s_listpair.append(pair);
-			}
-			lstWidget = _pShipWindow->findChildren<QWidget *>();
-			foreach(w, lstWidget)
-			{
-				QPair<QString, QWidget*> pair;
-				pair.first = w->styleSheet();
-				pair.second = w;
-				s_listpair.append(pair);
-			}
-
-			QPair<QString, QWidget*> p;
-			foreach(p, s_listpair)
-			{
-				w = p.second;
-				QString stylestr = w->styleSheet();
-				stylestr += "background-color: rgba(45, 45, 48, 30%);";
-				w->setStyleSheet(stylestr);
-			}
-		}
-
-		applyCss(QWebViewCSSIndex::Transparent);
-	}
-	else
-	{
-		QPair<QString, QWidget*> p;
-		if (!_bUseIE)
-		{
-			foreach(p, s_listpair)
-			{
-				if (isAncestorOf(p.second) || _pInfoWindow->isAncestorOf(p.second) || _pTimerWindow->isAncestorOf(p.second))
-				{
-					p.second->setStyleSheet(p.first);
-				}
-			}
-			s_listpair.clear();
-		}
-
-		applyCss(QWebViewCSSIndex::Normal);
-	}
-	/*
-	if (checked)
-	{
-		applyCss(QWebViewCSSIndex::Compact);
-	}
-	else
-	{
-		applyCss(QWebViewCSSIndex::Normal);
-	}
-	*/
 }
 
 void MainWindow::slotWebViewException(int code, const QString &source, const QString &desc, const QString &help)
@@ -607,25 +223,9 @@ void MainWindow::onGetNetworkReply(QNetworkReply * reply)
 	}
 }
 
-void MainWindow::setSleepMode(bool val)
-{
-	if (_bSleep != val)
-	{
-		_bSleep = val;
-
-		emit this->sigToggleSleepMode(_bSleep);
-	}
-//	qDebug("state changed");
-}
-
-bool MainWindow::isSleepMode()
-{
-	return _bSleep;
-}
-
 QWidget* MainWindow::getBrowserWidget()
 {
-	if (_bUseIE)
+	if (_webWidgetType == WebWidgetType::IE)
 	{
 		return _axWidget;
 	}
@@ -634,7 +234,7 @@ QWidget* MainWindow::getBrowserWidget()
 
 void MainWindow::navigateTo(const QString& urlString)
 {
-	if (_bUseIE)
+	if (_webWidgetType == WebWidgetType::IE)
 	{
 		_axWidget->Navigate(urlString);
 	}
@@ -646,7 +246,7 @@ void MainWindow::navigateTo(const QString& urlString)
 
 void MainWindow::navigateReload()
 {
-	if (_bUseIE)
+	if (_webWidgetType == WebWidgetType::IE)
 	{
 		rebuildIE(true);
 //		_axWidget->Refresh();
@@ -661,824 +261,7 @@ void MainWindow::navigateReload()
 
 bool MainWindow::isUsingIE()
 {
-	return _bUseIE;
-}
-
-void MainWindow::slotTogglePanicTimer(int timeVal)
-{
-	if (timeVal > 0)
-	{
-		_panicTimer->setTimerType(Qt::PreciseTimer);
-		_panicTimer->start(timeVal);
-	}
-	else
-	{
-		_panicTimer->stop();
-	}
-}
-
-
-void MainWindow::setWebSettings()
-{
-	QObject * exceptionSender = NULL;
-	if (_proxyMode == ProxyMode::Fid)
-	{
-		_pFid = new FidCOM::FidCOMClass(this);
-
-		loadCertKey();
-		_pFid->SetupCertificate(_certStr, _keyStr);
-		_pFid->InstallCertificate(_certStr, _keyStr);
-		saveCertKey();
-		exceptionSender = _pFid;
-
-		_pFid->SetBeforeRequest((int)&MainWindow::BeforeRequestFunc);
-		_pFid->SetAfterSessionComplete((int)&MainWindow::AfterSessionCompleteFunc);
-
-		_pFid->Startup(_useport, false, true);
-	}
-	else if (_proxyMode == ProxyMode::Nekoxy)
-	{
-		_pNekoxy = new Nekoxy::HttpProxy(this);
-		exceptionSender = _pNekoxy;
-
-		_pNekoxy->SetAfterSessionComplete((int)&MainWindow::AfterSessionCompleteFunc);
-
-		_pNekoxy->Startup(_useport, false, true);
-	}
-	else if (_proxyMode == ProxyMode::Titanium)
-	{
-		_pTitanium = new Titanium_Web_Proxy::ProxyServer(this);
-		exceptionSender = _pTitanium;
-		_pTitanium->SetMakecertPath(QApplication::applicationDirPath() + "/makecert.exe");
-
-		_pTitanium->SetAfterSessionComplete((int)&MainWindow::AfterSessionCompleteFunc);
-		_pTitanium->Startup(_useport);
-	}
-
-	if (_proxyMode != ProxyMode::NoProxy && _proxyMode != ProxyMode::QtProxy)
-	{
-		if (exceptionSender)
-		{
-			connect(
-				exceptionSender,
-				SIGNAL(exception(int, const QString &, const QString &, const QString &)),
-				this,
-				SLOT(slotWebViewException(int, const QString &, const QString &, const QString &)));
-		}
-		
-		QNetworkProxyFactorySet * proxies = new QNetworkProxyFactorySet();
-		proxies->init(_useport);
-//		QNetworkProxyFactory::setApplicationProxyFactory(proxies);
-		QNetworkProxy::setApplicationProxy(proxies->_proxyhttp);
-	}
-
-	if (!_bUseIE)
-	{
-		/*
-		KQWebPage * page = new KQWebPage();
-		_webView->setPage(page);
-
-		_webView->setContextMenuPolicy(Qt::PreventContextMenu);
-		/*
-		if (_proxyMode == ProxyMode::QtProxy)
-		{
-			KQNetworkAccessManager * manager = new KQNetworkAccessManager(this);
-			_webView->page()->setNetworkAccessManager(manager);
-		}
-		CookieJar* jar = new CookieJar(this);
-		_webView->page()->setCookieJar(jar);
-		
-		//WebView
-		QNetworkDiskCache *cache = new QNetworkDiskCache(this);
-		cache->setCacheDirectory(QStandardPaths::writableLocation(QStandardPaths::CacheLocation));
-		cache->setMaximumCacheSize(1073741824); //about 1024MB
-		_webView->page()->networkAccessManager()->setCache(cache);
-		
-		QDir dir(QStandardPaths::writableLocation(QStandardPaths::CacheLocation));
-		if(dir.exists()){
-			dir.removeRecursively();
-		}
-		*/
-
-		//    QNetworkProxyFactory::setUseSystemConfiguration(false);
-		//    updateProxyConfiguration();
-	}
-
-
-	auto webSettingFunc = [](QWebEngineSettings* websetting)
-	{
-		//JavaScript関連設定
-		websetting->setAttribute(QWebEngineSettings::PluginsEnabled, true);
-		websetting->setAttribute(QWebEngineSettings::JavascriptCanOpenWindows, true);
-		//		websetting->setAttribute(QWebEngineSettings::JavascriptCanCloseWindows, true);
-		websetting->setAttribute(QWebEngineSettings::PluginsEnabled, true);
-		websetting->setAttribute(QWebEngineSettings::JavascriptEnabled, true);
-		//フォント設定
-#if defined(Q_OS_WIN32)
-		//    websetting->setFontFamily(QWebSettings::StandardFont, "ＭＳ Ｐゴシック");
-		//    websetting->setFontFamily(QWebSettings::StandardFont, "MS PGothic");
-		websetting->setFontFamily(QWebEngineSettings::StandardFont, "Meiryo UI");
-		websetting->setFontFamily(QWebEngineSettings::SerifFont, "MS PMincho");
-		websetting->setFontFamily(QWebEngineSettings::SansSerifFont, "MS PGothic");
-		websetting->setFontFamily(QWebEngineSettings::FixedFont, "MS Gothic");
-#elif defined(Q_OS_LINUX)
-#elif defined(Q_OS_MAC)
-		websetting->setFontFamily(QWebEngineSettings::StandardFont, "ヒラギノ角ゴPro");
-		websetting->setFontFamily(QWebEngineSettings::SerifFont, "ヒラギノ明朝Pro");
-		websetting->setFontFamily(QWebEngineSettings::SansSerifFont, "ヒラギノ角ゴPro");
-		websetting->setFontFamily(QWebEngineSettings::FixedFont, "Osaka");
-#else
-#endif
-	};
-
-	webSettingFunc(QWebEngineSettings::defaultSettings());
-	if (!_bUseIE)
-	{
-		webSettingFunc(_webView->settings());
-		webSettingFunc(_webView->page()->settings());
-	}
-
-}
-
-void MainWindow::AdjustVolume(int vol)
-{
-	HRESULT hr = S_OK;
-
-	IMMDevice* pDevice = NULL;
-	IMMDeviceEnumerator* pEnumerator = NULL;
-	IAudioSessionManager2* pSessionManager = NULL;
-
-
-	// Create the device enumerator.
-	(hr = CoCreateInstance(
-		__uuidof(MMDeviceEnumerator),
-		NULL, CLSCTX_ALL,
-		__uuidof(IMMDeviceEnumerator),
-		(void**)&pEnumerator));
-
-	// Get the default audio device.
-	(hr = pEnumerator->GetDefaultAudioEndpoint(
-		eRender, eConsole, &pDevice));
-
-	// Get the session manager.
-	(hr = pDevice->Activate(
-		__uuidof(IAudioSessionManager2), CLSCTX_ALL,
-		NULL, (void**)&pSessionManager));
-
-	IAudioSessionControl *pControl;
-	pSessionManager->GetAudioSessionControl(NULL, 0, &pControl);
-	ISimpleAudioVolume *pSimpleVolume;
-	pSessionManager->GetSimpleAudioVolume(NULL, 0, &pSimpleVolume);
-	GUID nid = GUID_NULL;
-
-	if (vol < 0)
-	{
-		if (_bLowVol)
-		{
-			vol = 12;
-		}
-		else
-		{
-			vol = 100;
-		}
-	}
-
-	pSimpleVolume->SetMasterVolume(vol / 100.0f, &nid);
-
-	// Clean up.
-	SAFE_RELEASE(pSessionManager);
-	SAFE_RELEASE(pEnumerator);
-	SAFE_RELEASE(pDevice);
-}
-
-void MainWindow::slotSoundEnded()
-{
-	if (isMinimized())
-	{
-		AdjustVolume(0);
-	}
-	else
-	{
-		AdjustVolume(-1);
-	}
-}
-
-void MainWindow::showEvent(QShowEvent *event)
-{
-	MainWindowBase::showEvent(event);
-	if (!_pTaskbarButton)
-	{
-		_pTaskbarButton = new QWinTaskbarButton(this);
-		_pTaskbarButton->setWindow(windowHandle());
-		QWinTaskbarProgress * pTaskbarProgress = _pTaskbarButton->progress();
-		pTaskbarProgress->setVisible(true);
-
-		pTaskbarProgress->setMinimum(0);
-		pTaskbarProgress->setMaximum(10000);
-		pTaskbarProgress->setValue(0);
-		pTaskbarProgress->show();
-	}
-}
-
-void MainWindow::SetProgressBarPos(int pos, ProgressBarState state)
-{
-	if (_pTaskbarButton)
-	{
-		QWinTaskbarProgress * pProgress = _pTaskbarButton->progress();
-		_pTaskbarButton->progress()->setValue(pos);
-		switch (state)
-		{
-		case ProgressBarState::Normal:
-			if (pProgress->isPaused() || pProgress->isStopped())
-			{
-				pProgress->resume();
-			}
-			break;
-		case ProgressBarState::Paused:
-			if (pProgress->isStopped())
-			{
-				pProgress->resume();
-			}
-			if (!pProgress->isPaused())
-			{
-				pProgress->setPaused(true);
-			}
-			break;
-		case ProgressBarState::Stopped:
-			if (pProgress->isPaused())
-			{
-				pProgress->resume();
-			}
-			if (!pProgress->isStopped())
-			{
-				pProgress->stop();
-			}
-			break;
-		}
-	}
-}
-
-void MainWindow::on_pbMoveTogether_toggled(bool checked)
-{
-	_bMoveSubTogether = checked;
-}
-
-void MainWindow::on_pbRefresh_clicked()
-{
-	QMessageBox * pMessageBox = new QMessageBox(
-		QMessageBox::NoIcon
-		, QString::fromLocal8Bit("")
-		, QString::fromLocal8Bit("リロードしますか？")
-		, QMessageBox::Yes | QMessageBox::No
-		, this
-		, Qt::Dialog | Qt::MSWindowsFixedSizeDialogHint | Qt::FramelessWindowHint);
-	pMessageBox->setDefaultButton(QMessageBox::No);
-	pMessageBox->setAttribute(Qt::WA_DeleteOnClose, true);
-	int reply = pMessageBox->exec();
-
-	if (reply == QMessageBox::Yes)
-	{
-		navigateReload();
-	}
-
-}
-
-void MainWindow::on_pbScreenshot_clicked()
-{
-	shootScreen();
-}
-
-void MainWindow::onPanic()
-{
-	QString jsstr = QString::fromLocal8Bit("window.alert('%1'); ").arg(QString::fromLocal8Bit("パニック！"));
-	if (_bUseIE)
-	{
-		// TODO
-		QMessageBox* pMessageBox = new QMessageBox(
-			QMessageBox::NoIcon
-			, QString::fromLocal8Bit("")
-			, QString::fromLocal8Bit("パニック！")
-			, QMessageBox::Ok
-			, this
-			, Qt::Dialog | Qt::MSWindowsFixedSizeDialogHint | Qt::FramelessWindowHint);
-		pMessageBox->setDefaultButton(QMessageBox::NoButton);
-		pMessageBox->setAttribute(Qt::WA_DeleteOnClose, true);
-		pMessageBox->exec();
-	}
-	else
-	{
-		_webView->page()->runJavaScript(jsstr);
-	}
-	if (_panicTimer)
-	{
-		_panicTimer->stop();
-	}
-}
-
-void MainWindow::onDoJobFuel()
-{
-	auto& cm = ControlManager::getInstance();
-	cm.Terminate();
-
-	if (!(QApplication::queryKeyboardModifiers()&Qt::ShiftModifier))
-	{
-		QMessageBox* pMessageBox = new QMessageBox(
-			QMessageBox::NoIcon
-			, QString::fromLocal8Bit("")
-			, QString::fromLocal8Bit("終了条件")
-			, QMessageBox::NoButton
-			, this
-			, Qt::Dialog | Qt::MSWindowsFixedSizeDialogHint | Qt::FramelessWindowHint);
-
-		auto resetButton = pMessageBox->addButton(QString::fromLocal8Bit("クリア"), QMessageBox::ResetRole);
-		auto yusouButton = pMessageBox->addButton(QString::fromLocal8Bit("輸送×3"), QMessageBox::AcceptRole);
-		auto southEastButton = pMessageBox->addButton(QString::fromLocal8Bit("南西×5"), QMessageBox::ActionRole);
-		auto foreverButton = pMessageBox->addButton(QString::fromLocal8Bit("永久"), QMessageBox::RejectRole);
-		Q_UNUSED(foreverButton);
-
-		pMessageBox->setDefaultButton(QMessageBox::NoButton);
-//		pMessageBox->setAttribute(Qt::WA_DeleteOnClose, true);
-		pMessageBox->exec();
-
-		auto clickedButton = pMessageBox->clickedButton();
-		if (clickedButton == resetButton)
-		{
-			KanSaveData::getInstance().resetTotals();
-			KanDataConnector::getInstance().callUpdateOverviewTable();
-			return;
-		}
-		else if (clickedButton == yusouButton)
-		{
-			cm.setStopWhen(ControlManager::StopWhen::Yusou3);
-		}
-		else if (clickedButton == southEastButton)
-		{
-			cm.setStopWhen(ControlManager::StopWhen::SouthEast5);
-		}
-		delete pMessageBox;
-	}
-
-	if (cm.isFlagshipOnly(0))
-	{
-		cm.BuildNext_Fuel();
-	}
-	else
-	{
-		cm.BuildNext_SouthEast();
-	}
-
-	cm.StartJob();
-}
-
-void MainWindow::onDoJobKira()
-{
-	auto& cm = ControlManager::getInstance();
-	cm.Terminate();
-
-	cm.LoadToDoShipList_Kira();
-
-	if (QApplication::queryKeyboardModifiers()&Qt::ShiftModifier)
-	{
-		cm.BuildNext_Kira(true);
-	}
-	else
-	{
-		cm.BuildNext_Kira();
-	}
-	cm.StartJob();
-
-}
-
-void MainWindow::onDoJobLevel()
-{
-	auto& cm = ControlManager::getInstance();
-	cm.Terminate();
-
-	if (QApplication::queryKeyboardModifiers()&Qt::ShiftModifier)
-	{
-		cm.setLevelNoWait(false);
-	}
-	else
-	{
-		cm.setLevelNoWait(true);
-	}
-	if (cm.BuildNext_Level())
-	{
-		cm.StartJob();
-	}
-}
-
-void MainWindow::onDoJobExpedition(bool bDo)
-{
-	auto& cm = ControlManager::getInstance();
-	cm.Terminate();
-	if (bDo)
-	{
-		if (cm.BuildNext_Expedition())
-		{
-			cm.StartJob();
-		}
-		else
-		{
-			return;
-		}
-	}
-	ui->pbWaitExpedition->setChecked(bDo);
-}
-
-void MainWindow::onToggleJobExpedition()
-{
-	auto& cm = ControlManager::getInstance();
-	if (cm.isTerminated())
-	{
-		onDoJobExpedition(true);
-	}
-	else
-	{
-		onDoJobExpedition(false);
-	}
-}
-
-void MainWindow::onExportAllList()
-{
-	KanSaveData* pksd = &KanSaveData::getInstance();
-	KanDataConnector * pkdc = &KanDataConnector::getInstance();
-	struct ExportInfo
-	{
-		int shipno;
-		int shiptype;
-		int shipid;
-		QString shipname;
-		int lv;
-		int cond;
-	};
-	QList<ExportInfo> infos;
-	if (pksd->portdata.api_ship.size())
-	{
-		for (auto& ship:pksd->portdata.api_ship)
-		{
-			if (ship.api_lv < 10)
-			{
-				continue;
-			}
-
-			auto pmstship = pkdc->findMstShipFromShipid(ship.api_ship_id);
-			if (pmstship)
-			{
-				ExportInfo info;
-				info.shipno = ship.api_id;
-				info.shiptype = pmstship->api_stype;
-				info.shipid = ship.api_ship_id;
-				info.lv = ship.api_lv;
-				info.shipname = pmstship->api_name;
-				info.cond = ship.api_cond;
-				infos.append(info);
-			}
-		}
-	}
-	if (infos.size())
-	{
-		qSort(infos.begin(), infos.end(), [](const ExportInfo&left, const ExportInfo&right)
-		{
-			if (left.shiptype < right.shiptype)
-			{
-				return true;
-			}
-			else if (left.shiptype == right.shiptype)
-			{
-				/*
-				if (left.cond < right.cond)
-				{
-					return true;
-				}
-				else if (left.cond == right.cond)
-				{*/
-					if (left.shipid < right.shipid)
-					{
-						return true;
-					}
-					else if (left.shipid == right.shipid)
-					{
-						if (left.shipno < right.shipno)
-						{
-							return true;
-						}
-					}
-//				}
-			}
-			return false;
-		});
-
-
-		QFile * file = new QFile(QApplication::applicationDirPath() + "/action/" + "exported.table");
-		if (file)
-		{
-			if (file->open(QIODevice::WriteOnly | QIODevice::Text))
-			{
-				int lastShipType = -1;
-				for (auto& info:infos)
-				{
-					QString str;
-					if (lastShipType >= 0 && lastShipType != info.shiptype)
-					{
-						str += "\n";
-					}
-					lastShipType = info.shiptype;
-					str += "-\t" +
-						QString::number(info.shipno) + "\t" + 
-						info.shipname + "\t" + 
-						QString::number(info.lv) + "\t" +
-						QString::number(info.cond) + "\n";
-
-					QTextStream outstream(file);
-					outstream.setCodec("UTF-16");
-					outstream.setGenerateByteOrderMark(true);
-					outstream << str;
-				}
-
-				file->close();
-			}
-
-
-			delete file;
-		}
-	}
-	// export master data
-	KanSaveData::getInstance().logAllExport();
-}
-
-void MainWindow::onTerminateJob()
-{
-	ControlManager::getInstance().Terminate();
-}
-
-void MainWindow::slotSetLogAll(bool bSet)
-{
-	KanDataConnector::getInstance().setOutputAllLog(bSet);
-}
-
-void MainWindow::onJobPauseResume()
-{
-	if (ControlManager::getInstance().isPaused())
-	{
-		ControlManager::getInstance().Resume();
-	}
-	else
-	{
-		ControlManager::getInstance().Pause();
-	}
-}
-
-void MainWindow::onJobPauseNext()
-{
-	if (ControlManager::getInstance().isPaused())
-	{
-		ControlManager::getInstance().Resume();
-	}
-	else
-	{
-		ControlManager::getInstance().togglePauseNext();
-	}
-	ui->pbPauseNext->setChecked(ControlManager::getInstance().getPauseNextVal());
-}
-
-void MainWindow::on_pbSwitchScreenshot_toggled(bool checked)
-{
-	if (checked)
-	{
-		shootScreen();
-		_pScreenshotTimer->setTimerType(Qt::PreciseTimer);
-		_pScreenshotTimer->start(250);
-	}
-	else
-	{
-		_pScreenshotTimer->stop();
-	}
-}
-
-void MainWindow::shootScreen()
-{
-	if (ui->pbCheckTransparent->isChecked())
-	{
-		applyCss(QWebViewCSSIndex::Normal);
-	}
-	QPixmap pixmap = QPixmap::grabWidget(getBrowserWidget());
-
-	QString filename = QApplication::applicationDirPath();
-	filename += "/screenshot/";
-
-	filename += QDateTime::currentDateTime().toString("yyyy_MM_dd_HH_mm_ss_zzz.png");
-
-	pixmap.save(filename, "PNG");
-	if (ui->pbCheckTransparent->isChecked())
-	{
-		applyCss(QWebViewCSSIndex::Transparent);
-	}
-}
-
-void MainWindow::setupCss()
-{	
-	/*
-	 *
-body {
-	margin:0;
-	overflow:hidden
-}
-
-#game_frame {
-	position:fixed;
-	top:-16px;
-	left:-50px;
-	z-index:1
-}
-*/
-	_ieCsses[(int)QWebViewCSSIndex::Normal] = "\
-								body {\
-									margin:0;\
-									overflow:hidden;\
-								}\
-								#game_frame{\
-									position:fixed;\
-									top:-16px;\
-									left:-50px;\
-									z-index:1\
-								 }";
-	_webViewCsses[(int)QWebViewCSSIndex::Normal] = QUrl(QString("data:text/css;charset=utf-8;base64,") 
-		+ _ieCsses[(int)QWebViewCSSIndex::Normal].toUtf8().toBase64());
-//		(QUrl("data:text/css;charset=utf-8;base64,Ym9keSB7DQoJbWFyZ2luOjA7DQoJb3ZlcmZsb3c6aGlkZGVuDQp9DQoNCiNnYW1lX2ZyYW1lIHsNCglwb3NpdGlvbjpmaXhlZDsNCgl0b3A6LTE2cHg7DQoJbGVmdDotNTBweDsNCgl6LWluZGV4OjENCn0="));
-
-
-	/*
-	*
-	body {
-	margin:0;
-	overflow:hidden
-	}
-
-	#game_frame {
-	position:fixed;
-	top:-16px;
-	left:-50px;
-	z-index:1
-	zoom:0.5
-	}
-	*/
-	
-	_ieCsses[(int)QWebViewCSSIndex::Compact] = "\
-								body {\
-									margin:0;\
-									overflow:hidden;\
-								}\
-								#game_frame{\
-									position:fixed;\
-									top:-16px;\
-									left:-50px;\
-									z-index:1\
-																										 }";
-	_webViewCsses[(int)QWebViewCSSIndex::Compact] = QUrl(QString("data:text/css;charset=utf-8;base64,")
-		+ _ieCsses[(int)QWebViewCSSIndex::Compact].toUtf8().toBase64());
-	/*
-body {
-	margin:0;
-	overflow:hidden;
-	opacity: 0.4;
-}
-
-#game_frame {
-	position:fixed;
-	top:-16px;
-	left:-50px;
-	z-index:1;
-}
-*/
-//	csses[QWebViewCSS::TRANSPARENT] = QUrl("data:text/css;charset=utf-8;base64,Ym9keSB7DQoJbWFyZ2luOjA7DQoJb3ZlcmZsb3c6aGlkZGVuOw0KCW9wYWNpdHk6IDAuNDsNCn0NCg0KI2dhbWVfZnJhbWUgew0KCXBvc2l0aW9uOmZpeGVkOw0KCXRvcDotMTZweDsNCglsZWZ0Oi01MHB4Ow0KCXotaW5kZXg6MTsNCn0=");
-
-	/*
-	body {
-	margin:0;
-	overflow:hidden;
-	opacity: 0.2;
-	}
-
-	#game_frame {
-	position:fixed;
-	top:-16px;
-	left:-50px;
-	z-index:1;
-	}
-	*/
-	_webViewCsses[(int)QWebViewCSSIndex::Transparent] = 
-		QUrl("data:text/css;charset=utf-8;base64,Ym9keSB7DQoJbWFyZ2luOjA7DQoJb3ZlcmZsb3c6aGlkZGVuOw0KCW9wYWNpdHk6IDAuMjsNCn0NCg0KI2dhbWVfZnJhbWUgew0KCXBvc2l0aW9uOmZpeGVkOw0KCXRvcDotMTZweDsNCglsZWZ0Oi01MHB4Ow0KCXotaW5kZXg6MTsNCn0=");
-
-	// do not set opacity
-	_ieCsses[(int)QWebViewCSSIndex::Transparent] = "\
-										body {\
-											margin:0;\
-											overflow:visible;\
-										}\
-										#game_frame{\
-											position:fixed;\
-											top:-16px;\
-											left:-50px;\
-											z - index:1;\
-										}\
-										";
-	
-	applyCss(QWebViewCSSIndex::Normal);
-}
-
-void MainWindow::loadSettings()
-{
-	QString filename = QApplication::applicationDirPath();
-	filename += "/settings.ini";
-	QSettings* setting = new QSettings(filename, QSettings::IniFormat);
-	setting->setIniCodec("UTF-8");
-
-	const QString itemUseIE = "UseIE";
-	const QString itemGameUrl = "GameUrl";
-	const QString itemUrlId = "UrlId";
-	const QString itemUsePort = "UsePort";
-	const QString itemIntervalMul = "IntervalMul";
-	const QString itemProxyMode = "ProxyMode";
-	const QString itemExpeditionMode = "ExpeditionMode";
-	
-	setting->beginGroup("Settings");
-	if (!setting->contains(itemUseIE))
-	{
-		if (QSysInfo::WindowsVersion >= QSysInfo::WV_WINDOWS8)
-		{
-			setting->setValue(itemUseIE, true);
-		}
-		else
-		{
-			setting->setValue(itemUseIE, false);
-		}
-	}
-	if (!setting->contains(itemGameUrl))
-	{
-		setting->setValue(itemGameUrl, "http://www.dmm.com/netgame/social/-/gadgets/=/app_id=854854/?f1=1&p1=0");
-	}
-	if (!setting->contains(itemUrlId))
-	{
-		setting->setValue(itemUrlId, "app_id=854854");
-	}
-	if (!setting->contains(itemUsePort))
-	{
-		setting->setValue(itemUsePort, 13347);
-	}
-	if (!setting->contains(itemIntervalMul))
-	{
-		if (QSysInfo::WindowsVersion >= QSysInfo::WV_WINDOWS8)
-		{
-			setting->setValue(itemIntervalMul, 150);
-		}
-		else
-		{
-			setting->setValue(itemIntervalMul, 100);
-		}
-	}
-	if (!setting->contains(itemProxyMode))
-	{
-		setting->setValue(itemProxyMode, "Nekoxy");
-	}
-	if (!setting->contains(itemExpeditionMode))
-	{
-		setting->setValue(itemExpeditionMode, "General");
-	}
-
-	_bUseIE = setting->value(itemUseIE).toBool();
-	_gameUrl = setting->value(itemGameUrl).toString();
-	_gameUrlId = setting->value(itemUrlId).toString();
-	_useport = setting->value(itemUsePort).toInt();
-	ControlManager::getInstance().setIntervalMul(setting->value(itemIntervalMul).toInt()/100.0);
-
-	QString proxymode = setting->value(itemProxyMode).toString();
-	if (!proxymode.compare("NoProxy", Qt::CaseInsensitive))
-	{
-		_proxyMode = ProxyMode::NoProxy;
-	}
-	else if (!proxymode.compare("QtProxy", Qt::CaseInsensitive))
-	{
-		_proxyMode = ProxyMode::QtProxy;
-	}
-	else if (!proxymode.compare("Fid", Qt::CaseInsensitive))
-	{
-		_proxyMode = ProxyMode::Fid;
-	}
-	else if (!proxymode.compare("Nekoxy", Qt::CaseInsensitive))
-	{
-		_proxyMode = ProxyMode::Nekoxy;
-	}
-	else if (!proxymode.compare("Titanium", Qt::CaseInsensitive))
-	{
-		_proxyMode = ProxyMode::Titanium;
-	}
-
-	ExpeditionManager::getInstance().BuildByPreset(setting->value(itemExpeditionMode).toString());
-
-	setting->endGroup();
-
-	delete setting;
+	return _webWidgetType == WebWidgetType::IE;
 }
 
 void MainWindow::rebuildIE(bool bNavigate)
@@ -1510,87 +293,6 @@ void MainWindow::rebuildIE(bool bNavigate)
 	}
 }
 
-void MainWindow::loadCertKey()
-{
-	QSettings settings("h5nc", "KanPlay");
-	_certStr = settings.value("CertStr", "").toString();
-	_keyStr = settings.value("KeyStr", "").toString();
-}
-
-void MainWindow::saveCertKey()
-{
-	QSettings settings("h5nc", "KanPlay");
-	settings.setValue("CertStr", _certStr);
-	settings.setValue("KeyStr", _keyStr);
-}
-
-void MainWindow::applyCss(QWebViewCSSIndex css)
-{
-	if (css < QWebViewCSSIndex::Normal || css >= QWebViewCSSIndex::Max)
-	{
-		return;
-	}
-	if (_bUseIE)
-	{
-		_applyCssWhenLoaded = css;
-		if (!_bIEPageLoaded)
-		{
-			return;
-		}
-
-		IDispatch* htmlDoc = _axWidget->Document();
-		if (!htmlDoc)
-		{
-			return;
-		}
-		QAxObject * htmlDocObj = new QAxObject(htmlDoc);
-
-		if (!htmlDocObj)
-		{
-			return;
-		}
-
-		QAxObject * styleSheetObj = htmlDocObj->querySubObject("createStyleSheet()");
-		if (!styleSheetObj)
-		{
-			return;
-		}
-
-		styleSheetObj->dynamicCall("setCssText(QString)", _ieCsses[(int)css]);
-		_applyCssWhenLoaded = QWebViewCSSIndex::Invalid;
-		delete htmlDocObj;
-	}
-	else
-	{
-		QString jsStr = "function addStyleString(str) {\
-			var node = document.createElement('style');\
-		node.innerHTML = str;\
-		document.body.appendChild(node);\
-	}\
-	addStyleString('body {margin:0;overflow:hidden;}');\
-	addStyleString('#game_frame{position:fixed;top:-16px;left:-50px;z-index:1}'); \
-	addStyleString('body { background: silver }'); ";
-		_webView->page()->runJavaScript(jsStr);
-//		_webView->page()->settings()->setUserStyleSheetUrl(_webViewCsses[(int)css]);
-	}
-}
-
-void MainWindow::slotScreenshotTimeout()
-{
-	shootScreen();
-}
-
-void MainWindow::slotJobtTimeout()
-{
-	ControlManager::getInstance().Run();
-}
-
-void MainWindow::on_pbCheckLowVol_toggled(bool checked)
-{
-	_bLowVol = checked;
-	AdjustVolume(-1);
-}
-
 void MainWindow::slotNavigateComplete2(IDispatch*, QVariant& url)
 {
 	// warning : does not work with 5.6
@@ -1605,83 +307,4 @@ void MainWindow::slotNavigateComplete2(IDispatch*, QVariant& url)
 	{
 		applyCss(_applyCssWhenLoaded);
 	}
-}
-
-void MainWindow::setPauseNextChanged(bool bVal)
-{
-	ui->pbPauseNext->setChecked(bVal);
-}
-
-void MainWindow::setJobTerminated()
-{
-	ui->pbPauseNext->setChecked(false);
-	ui->pbWaitExpedition->setChecked(false);
-}
-
-void MainWindow::onIncreaseSouthEast()
-{
-	KanSaveData::getInstance().adjustSouthEast(1);
-	KanDataConnector::getInstance().callUpdateOverviewTable();
-}
-
-void MainWindow::onDecreaseSouthEast()
-{
-	KanSaveData::getInstance().adjustSouthEast(-1);
-	KanDataConnector::getInstance().callUpdateOverviewTable();
-}
-
-QWindowsEventFilter::QWindowsEventFilter()
-{
-
-}
-
-bool QWindowsEventFilter::nativeEventFilter(const QByteArray &eventType, void *message, long *result)
-{
-	Q_UNUSED(result);
-	if (eventType == "windows_generic_MSG")
-	{
-		MSG *msg = static_cast<MSG *>(message);
-		if (msg->message == WM_POWERBROADCAST)
-		{
-			auto mainWindow = MainWindow::mainWindow();
-			if (!mainWindow)
-			{
-				return false;
-			}
-			if (msg->wParam == PBT_APMSUSPEND)
-			{
-				// sleep
-				mainWindow->setSleepMode(true);
-				return false;
-			}
-			if (msg->wParam == PBT_APMRESUMEAUTOMATIC)
-			{
-				// resume
-				mainWindow->setSleepMode(false);
-				return false;
-			}
-			if (msg->wParam == PBT_POWERSETTINGCHANGE)
-			{
-				POWERBROADCAST_SETTING *ps = (POWERBROADCAST_SETTING*)(msg->lParam);
-				if (ps->PowerSetting == GUID_CONSOLE_DISPLAY_STATE)
-				{
-					DWORD dw = *(DWORD*)(ps->Data);
-					if (dw == 0x0)
-					{
-						// off
-//						mainWindow->setSleepMode(true);
-						return false;
-					}
-					if (dw == 0x1)
-					{
-						// on
-//						mainWindow->setSleepMode(false);
-						return false;
-					}
-				}
-			}
-			return false;
-		}
-	}
-	return false;
 }
