+++ mainwindow.h
@@ -1,33 +1,34 @@
 ﻿#ifndef MAINWINDOW_H
 #define MAINWINDOW_H
 
+#include <QtGlobal>
 #include <QMainWindow>
 #include "mainwindowbase.h"
 #include "infomainwindow.h"
 #include "timermainwindow.h"
 #include "weaponmainwindow.h"
 #include "shipmainwindow.h"
+
 #include <QShowEvent>
 #include <QWinTaskbarButton>
 #include <QNetworkReply>
+
+#if (QT_VERSION >= QT_VERSION_CHECK(5, 6, 0))
 #include <QWebEngineView>
+#else
+#include <QWebView>
+#endif
+
 #include <QAxWidget>
 #include "shdocvw.h"
 
 #include "fidcom.h"
 #include "nekoxy.h"
 #include "titanium_web_proxy.h"
+#include "qwindowseventfilter.h"
 
 using namespace SHDocVw;
 
-#include <QAbstractNativeEventFilter>
-
-class QWindowsEventFilter : public QAbstractNativeEventFilter
-{
-public:
-	QWindowsEventFilter();
-	virtual bool nativeEventFilter(const QByteArray &eventType, void *message, long *result);
-};
 enum class ProgressBarState
 {
 	Normal,
@@ -54,6 +55,20 @@ enum class ProxyMode
 	Titanium,
 };
 
+enum class WebWidgetType
+{
+	Webkit,
+	WebEngine,	// Qt5.6
+	IE,
+};
+
+enum class PlatformType
+{
+	Windows7,
+	SlowTablet,		// windows8
+	FastTablet,		// surface
+	Mac,
+};
 
 namespace Ui {
 class MainWindow;
@@ -66,6 +81,7 @@ class MainWindow : public MainWindowBase
 public:
 	explicit MainWindow(QWidget *parent = 0);
 	~MainWindow();
+
 
 	static inline MainWindow * mainWindow(){return s_pMainWindow;}
 	static inline void setMainWindow(MainWindow * pWindow){s_pMainWindow = pWindow;}
@@ -92,6 +108,9 @@ public:
 
 	void setPauseNextChanged(bool bVal);
 	void setJobTerminated();
+
+	WebWidgetType getWebWidgetType();
+	PlatformType getPlatformType();
 
 signals:
 	void sigParse(const QString &PathAndQuery, const QString &requestBody, const QString &responseBody);
@@ -195,16 +214,23 @@ private:
 
 	QWindowsEventFilter _windowsEventfilter;
 
+#if (QT_VERSION >= QT_VERSION_CHECK(5, 6, 0))
 	QWebEngineView* _webView = NULL;
+#else
+	QWebView* _webView = NULL;
+#endif
 	WebBrowser* _axWidget = NULL;
 
 	// settings
-	bool _bUseIE = false;
 	QString _gameUrl;
 	QString _gameUrlId;
 
 	QString _certStr;
 	QString _keyStr;
+
+	// need to be set
+	WebWidgetType _webWidgetType = WebWidgetType::Webkit;
+	PlatformType _platformType = PlatformType::Windows7;
 };
