+++ qwindowseventfilter.h
@@ -1,0 +1,13 @@
+#ifndef QWINDOWSEVENTFILTER_H
+#define QWINDOWSEVENTFILTER_H
+
+#include <QAbstractNativeEventFilter>
+
+class QWindowsEventFilter : public QAbstractNativeEventFilter
+{
+public:
+	QWindowsEventFilter();
+	virtual bool nativeEventFilter(const QByteArray &eventType, void *message, long *result);
+};
+
+#endif // QWINDOWSEVENTFILTER_H
